Reduced log level of experimental/local opt codes
RFC 6891 identifies the range of OPT codes from 65001 to 65534 as used
for experimental/local purposes. Receipt of answers containing these
OPT codes is not cause for alarm, and can be safely ignored. Reducing
logging of the receipt of these answers to FINE instead of WARNING.
Other unrecognized codes outside this range will continue to WARN.
Fixes #12

Signed-off-by: Andy Lintner <dev@beowulfe.com> (github: beowulfe)